Add files via upload
Included a Cipher Function to the code along with translational alphabet to within the GUI for DM reference. Including the .ttf of the cipher language I'm using (Standard Galactic Alphabet) as well as the first GUI attempt at using the Cipher prior to implementing it into my main GUI program.