Add files via upload
updated from Random_Generator_Monster.py This now includes 'Any' enhancements to CR functions so monsters are more readily found. Also I accidentally saved the the first Cipher update on top of the original Monster update version on my GitHub lol, whoops. So now this iteration is called Cipher. Point of note, if you want to successfully use the cipher option you will need to save the .ttf file i've included in this folder, install it, and save it within the same folder level as this project in order for it to read. I've gotten some feedback that when run on another computer, the dictonary housed within the Cipher Tab defaults to "emojis" of various kinds but when the function is ran with the ttf installed and properly stored, the generated image does use the SGA language. Working on that bug now, any comments on how to fix that are welcome.